--- src/worker_message.rs
@@ -1,17 +1,0 @@
-use super::*;
-
-#[derive(Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub(crate) enum WorkerMessage {
-  Clear,
-  DecibelRange { min: f32, max: f32 },
-  Done,
-  Error(String),
-  OscillatorFrequency(f32),
-  OscillatorGain(f32),
-  Record,
-  Render(Filter),
-  Resolution(u32),
-  Save,
-  Widget { name: String, widget: Widget },
-}
